requirements.txt
# Core data science libraries
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Machine learning libraries
xgboost>=1.7.0
lightgbm>=3.3.0
catboost>=1.2.0

# Deep learning
tensorflow>=2.13.0
keras>=2.13.0
torch>=2.0.0
torchvision>=0.15.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
bokeh>=3.2.0

# Data processing and feature engineering
feature-engine>=1.6.0
category-encoders>=2.6.0
imbalanced-learn>=0.11.0

# Geospatial and time series
geopandas>=0.13.0
folium>=0.14.0
contextily>=1.3.0
statsmodels>=0.14.0
pmdarima>=2.0.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.65.0
joblib>=1.3.0

# Jupyter and development
jupyter>=1.0.0
jupyterlab>=4.0.0
notebook>=6.5.0
ipywidgets>=8.0.0

# Data quality and validation
great-expectations>=0.17.0
pandera>=0.17.0

# Hyperparameter optimization
optuna>=3.3.0
hyperopt>=0.2.7
scikit-optimize>=0.9.0

# Model interpretation
shap>=0.42.0
lime>=0.2.0
eli5>=0.13.0

# Logging and monitoring
wandb>=0.15.0
mlflow>=2.5.0
loguru>=0.7.0

# API and web framework (for deployment)
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.25.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0
